finished with working on the importing to database
