+++ spm_backend/role/role.sql
@@ -1,7 +1,6 @@
 CREATE TABLE role (
    role_name  VARCHAR(20) NOT NULL,
-   skill_name VARCHAR(50) NOT NULL,
-   description VARCHAR(50) NOT NULL,
-   department VARCHAR(50) NOT NULL,
-   PRIMARY KEY (role_name, skill_name)
+   role_description TEXT NOT NULL,
+   
+   PRIMARY KEY (role_name)
 );
