--- spm_backend/role/role_skill.csv
@@ -1,3 +1,0 @@
-Role,Skill,Description,Department
-Software Developer,Software Development,Programme Software Products,Engineering
-
